add PathAnnotation class
