+++ src/main/kotlin/com/viartemev/requestmapper/annotations/jaxrs/JaxRsMappingAnnotation.kt
@@ -1,18 +1,11 @@
 package com.viartemev.requestmapper.annotations.jaxrs
 
 import com.intellij.psi.PsiAnnotation
-import com.intellij.psi.PsiAnnotationMemberValue
-import com.intellij.psi.PsiArrayInitializerMemberValue
 import com.intellij.psi.PsiMethod
-import com.intellij.psi.PsiReferenceExpression
 import com.viartemev.requestmapper.RequestMappingItem
 import com.viartemev.requestmapper.annotations.MappingAnnotation
-import com.viartemev.requestmapper.annotations.spring.extraction.BasePsiAnnotationValueVisitor
-import com.viartemev.requestmapper.annotations.spring.extraction.PsiAnnotationMemberValueExtractor
-import com.viartemev.requestmapper.annotations.spring.extraction.PsiArrayInitializerMemberValueExtractor
-import com.viartemev.requestmapper.annotations.spring.extraction.PsiReferenceExpressionExtractor
+import com.viartemev.requestmapper.annotations.PathAnnotation
 import com.viartemev.requestmapper.utils.fetchAnnotatedMethod
-import com.viartemev.requestmapper.utils.unquote
 
 abstract class JaxRsMappingAnnotation(val psiAnnotation: PsiAnnotation) : MappingAnnotation {
 
@@ -34,7 +27,7 @@ abstract class JaxRsMappingAnnotation(val psiAnnotation: PsiAnnotation) : Mappin
             ?.modifierList
             ?.annotations
             ?.filter { it.qualifiedName == PATH_ANNOTATION }
-            ?.flatMap { fetchMappingsFromAnnotation(it, ATTRIBUTE_NAME) }
+            ?.flatMap { PathAnnotation(it).fetchMappings(ATTRIBUTE_NAME) }
             ?.first() ?: ""
     }
 
@@ -43,21 +36,8 @@ abstract class JaxRsMappingAnnotation(val psiAnnotation: PsiAnnotation) : Mappin
             .modifierList
             .annotations
             .filter { it.qualifiedName == PATH_ANNOTATION }
-            .flatMap { fetchMappingsFromAnnotation(it, ATTRIBUTE_NAME) }
+            .flatMap { PathAnnotation(it).fetchMappings(ATTRIBUTE_NAME) }
             .firstOrNull() ?: ""
-    }
-
-    private fun fetchMappingsFromAnnotation(annotation: PsiAnnotation, parameter: String): List<String> {
-        return object : BasePsiAnnotationValueVisitor() {
-            override fun visitPsiArrayInitializerMemberValue(arrayAValue: PsiArrayInitializerMemberValue) =
-                PsiArrayInitializerMemberValueExtractor().extract(arrayAValue)
-
-            override fun visitPsiReferenceExpression(expression: PsiReferenceExpression) =
-                PsiReferenceExpressionExtractor().extract(expression)
-
-            override fun visitPsiAnnotationMemberValue(value: PsiAnnotationMemberValue) =
-                PsiAnnotationMemberValueExtractor().extract(value)
-        }.visit(annotation, parameter)
     }
 
     companion object {
